+++ src/main.c
@@ -6,7 +6,7 @@ int main(void)
 {
     Chunk* chunk = chunk_new();
 
-    size_t constant = chunk_add_constant(chunk, 1.2);
+    uint8_t constant = (uint8_t) chunk_add_constant(chunk, 1.2);
     chunk_write(chunk, OP_CODE_CONSTANT);
     chunk_write(chunk, constant);
